requirements.txt
langchain-community
streamlit
PyPDF2
python-dotenv
openai
faiss-cpu
